@@ -1,16 +1,20 @@
 {
   "name": "virtual-quiz",
-  "version": "0.0.0",
+  "version": "1.0.0",
   "license": "MIT",
   "scripts": {
     "ng": "ng",
     "start": "ng serve",
-    "build": "ng build --prod",
+    "build": "npm-run-all client:build server:build",
+    "client:build": "ng build --prod",
     "test": "ng test",
     "lint": "ng lint",
     "e2e": "ng e2e",
-    "server": "nodemon server/index.js",
-    "postinstall": "ng build --prod"
+    "server": "npm-run-all server:build server:watchAndRun",
+    "server:watchAndRun": "npm-run-all -r --parallel server:watch server:run",
+    "server:build": "tsc -p ./server",
+    "server:watch": "tsc -p server/tsconfig.dev.json",
+    "server:run": "nodemon --watch dist/server dist/server/server.js"
   },
   "engines": {
     "node": "^10.15.3",
@@ -18,37 +22,42 @@
   },
   "private": true,
   "dependencies": {
-    "@angular/animations": "~7.2.0",
-    "@angular/cdk": "~7.3.6",
-    "@angular/common": "~7.2.0",
-    "@angular/compiler": "~7.2.0",
-    "@angular/core": "~7.2.0",
-    "@angular/forms": "~7.2.0",
-    "@angular/http": "~7.2.0",
-    "@angular/material": "~7.3.6",
-    "@angular/platform-browser": "~7.2.0",
-    "@angular/platform-browser-dynamic": "~7.2.0",
-    "@angular/router": "~7.2.0",
+    "@angular/animations": "~8.0.2",
+    "@angular/cdk": "~8.0.1",
+    "@angular/common": "~8.0.2",
+    "@angular/compiler": "~8.0.2",
+    "@angular/core": "~8.0.2",
+    "@angular/forms": "~8.0.2",
+    "@angular/material": "~8.0.1",
+    "@angular/platform-browser": "~8.0.2",
+    "@angular/platform-browser-dynamic": "~8.0.2",
+    "@angular/router": "~8.0.2",
+    "body-parser": "~1.18.3",
     "core-js": "^2.5.4",
-    "express": "^4.16.4",
-    "expressjs": "^1.0.1",
+    "express": "^4.17.0",
+    "jsonwebtoken": "^8.5.1",
     "moment": "^2.24.0",
     "moment-timezone": "^0.5.23",
-    "mongodb": "^3.2.2",
-    "rxjs": "~6.4.0",
+    "mongodb": "^3.2.7",
+    "papaparse": "^5.0.0",
+    "rxjs": "~6.5.2",
     "tslib": "^1.9.0",
-    "web-animations-js": "^2.3.1",
-    "zone.js": "~0.8.26"
+    "web-animations-js": "github:angular/web-animations-js#release_pr208",
+    "zone.js": "~0.9.1"
   },
   "devDependencies": {
-    "@angular-devkit/build-angular": "~0.13.0",
-    "@angular/cli": "~7.3.6",
-    "@angular/compiler-cli": "~7.2.0",
-    "@angular/language-service": "~7.2.0",
+    "@angular-devkit/build-angular": "~0.800.0",
+    "@angular/cli": "~8.0.4",
+    "@angular/compiler-cli": "~8.0.2",
+    "@angular/language-service": "~8.0.2",
+    "@types/express": "^4.17.0",
     "@types/jasmine": "~2.8.8",
     "@types/jasminewd2": "~2.0.3",
+    "@types/jsonwebtoken": "^8.3.2",
+    "@types/moment-timezone": "^0.5.12",
+    "@types/mongodb": "^3.1.28",
     "@types/node": "~8.9.4",
-    "codelyzer": "~4.5.0",
+    "codelyzer": "^5.0.1",
     "jasmine-core": "~2.99.1",
     "jasmine-spec-reporter": "~4.2.1",
     "karma": "~4.0.0",
@@ -57,9 +66,11 @@
     "karma-jasmine": "~1.1.2",
     "karma-jasmine-html-reporter": "^0.2.2",
     "nodemon": "^1.18.10",
+    "npm-run-all": "^4.1.5",
     "protractor": "~5.4.0",
+    "request-stats": "^3.0.0",
     "ts-node": "~7.0.0",
     "tslint": "~5.11.0",
-    "typescript": "~3.2.2"
+    "typescript": "~3.4.5"
   }
 }
